+++ src/main/java/com/empresa/demo/model/Usuario.java
@@ -1,10 +1,13 @@
 package com.empresa.demo.model;
 
 import java.io.Serializable;
+import java.util.ArrayList;
 import java.util.Date;
 import java.util.HashSet;
+import java.util.List;
 import java.util.Set;
 
+import javax.persistence.CascadeType;
 import javax.persistence.Column;
 import javax.persistence.Entity;
 import javax.persistence.FetchType;
@@ -41,9 +44,16 @@ public class Usuario implements Serializable {
 
 	@Column(name = "fecha_registro")
 	private Date fecha_registro;
-	
-	@OneToMany(fetch = FetchType.EAGER,mappedBy = "username")
-	private Set<Rol> userRole=new HashSet<Rol>();
+
+	@OneToMany(fetch = FetchType.EAGER, mappedBy = "username")
+	private Set<Rol> userRole = new HashSet<Rol>();
+
+	// relación de uno a muchos hacia el comprobante, el objeto mapeado es "usuario"
+	// esta apuntado hacia el comprobante
+	// este nos creara un foreign key automaticamente y asi relacionando cliente y
+	// comprobante
+	@OneToMany(mappedBy = "usuario", fetch = FetchType.LAZY, cascade = CascadeType.ALL)
+	private List<Comprobante> comprobantes;
 
 	// metodo que nos ayudara a guardar la "fecha_registro" automaticamente,
 	// agarrara la fecha y hora actual
@@ -52,7 +62,9 @@ public class Usuario implements Serializable {
 		fecha_registro = new Date();
 	}
 
+	// inicializamos la lista en un arraylist
 	public Usuario() {
+		comprobantes = new ArrayList<Comprobante>();
 
 	}
 
@@ -144,6 +156,19 @@ public class Usuario implements Serializable {
 
 	public void setUserRole(Set<Rol> userRole) {
 		this.userRole = userRole;
+	}
+
+	public List<Comprobante> getComprobantes() {
+		return comprobantes;
+	}
+
+	public void setComprobantes(List<Comprobante> comprobantes) {
+		this.comprobantes = comprobantes;
+	}
+
+	// este metodo nos permitira agregar un comprobante a la vez
+	public void agregarComprobante(Comprobante comprobante) {
+		comprobantes.add(comprobante);
 	}
 
 	private static final long serialVersionUID = 1L;
