+++ src/main/resources/application.properties
@@ -11,7 +11,7 @@ server.port=8085
 spring.datasource.url=jdbc:mysql://localhost:3306/serviexpress?serverTimezone=America/Bogota
 
 spring.datasource.username=root
-spring.datasource.password=mysql
+spring.datasource.password=12345678
 
 spring.jpa.open-in-view=false
 #jpa hibernate
