+++ src/main/resources/templates/ver_detalle.html
@@ -1,0 +1,65 @@
+<!DOCTYPE html>
+<html xmlns:th="http://www.thymeleaf.org">
+<head>
+<link
+	href="https://cdn.jsdelivr.net/npm/bootstrap@5.0.2/dist/css/bootstrap.min.css"
+	rel="stylesheet"
+	integrity="sha384-EVSTQN3/azprG1Anm3QDgpJLIm9Nao0Yz1ztcQTwFspd3yD65VohhpuuCOmLASjC"
+	crossorigin="anonymous">
+<meta charset="UTF-8" />
+</head>
+<body>
+
+
+
+	<div class="container py-4">
+		<div class="card bg-ligth">
+
+			<div class="card-header" th:text="${titulo}"></div>
+			<div class="card-body"></div>
+
+
+			<ul class="list-group">
+				<li class="list-group-item"
+					th:text=" 'Usuario:  ' +${usuario.username}"></li>
+				<li class="list-group-item"
+					th:text=" 'Celular:  ' +${usuario.celular}"></li>
+				<li class="list-group-item"
+					th:text=" 'Documento de identidad:  ' +${usuario.documento_identidad}">
+				</li>
+			</ul>
+
+			<table>
+				<thead>
+					<tr>
+						<th>Comprobante</th>
+						<th>Fecha de Emisión</th>
+						<th>Tipo de documento</th>
+						<th>Metodo de Pago</th>
+						<th>Subtotal</th>
+						<th>IGV</th>
+						<th>Total</th>
+					</tr>
+				</thead>
+				<tbody>
+					<tr>
+						<td th:each="comprobante : ${usuario.comprobantes}"></td>
+						<td th:text="${comprobante.id_comprobante}"></td>
+						<td th:text="${comprobante.fecha_emision}"></td>
+						<td th:text="${comprobante.tipo_documento}"></td>
+						<td th:text="${comprobante.metodo_pago}"></td>
+						<td th:text="${comprobante.subtotal}"></td>
+						<td th:text="${comprobante.igv}"></td>
+						<td th:text="${comprobante.total}"></td>
+					</tr>
+				</tbody>
+			</table>
+
+
+
+
+		</div>
+	</div>
+
+</body>
+</html>
